+++ ncov19_dash/__init__.py
@@ -1,5 +1,7 @@
 from .dash_app import app
 from .flask_server import server
+from .cache import server_cache
+
 # from ncov19_dash.app import create_app
 
 # app = create_app()
