+++ ncov19_dash/components/__init__.py
@@ -1,10 +1,9 @@
-from .confirmed_cases_chart import confirmed_cases_chart
+# from .confirmed_cases_chart import confirmed_cases_chart
 from .daily_stats import daily_stats
 from .new_infection_trajectory_chart import new_infection_trajectory_chart
 from .news_feed import news_feed
 from .scatter_mapbox import confirmed_scatter_mapbox, drive_thru_scatter_mapbox
 from .twitter_feed import twitter_feed
-from .daily_stats import daily_stats
 from .daily_stats_mobile import daily_stats_mobile
 from .column_stats import last_updated
 from .column_stats_mobile import mobile_last_updated
