+++ ncov19_dash/pages/desktop.py
@@ -1,24 +1,10 @@
 import dash_bootstrap_components as dbc
 import dash_core_components as dcc
 import dash_html_components as html
-import dash_table
-import dash
-from dash_table.Format import Format, Scheme
-import dash_table.FormatTemplate as FormatTemplate
-from dash.dependencies import Input, Output, State
-from dash.exceptions import PreventUpdate
 
-from ncov19_dash.dash_app import app
-from ncov19_dash.utils.settings import STATES_COORD, REVERSE_STATES_MAP, STATE_LABELS
-
-from ncov19_dash.components import daily_stats
-from ncov19_dash.components import news_feed, twitter_feed
-from ncov19_dash.components import confirmed_cases_chart, new_infection_trajectory_chart
-from ncov19_dash.components import confirmed_scatter_mapbox, drive_thru_scatter_mapbox
-from ncov19_dash.components import last_updated
 from ncov19_dash.components import cases_chart, deaths_chart
-from ncov19_dash.components import stats_table
-from ncov19_dash.components.column_stats import STATES
+from ncov19_dash.components import last_updated
+from ncov19_dash.utils.settings import STATES_COORD, REVERSE_STATES_MAP, STATE_LABELS
 
 
 ################ TABS STYLING ####################
@@ -83,25 +69,6 @@ feed_tabs = dbc.Card(
 )
 
 
-#################### FEED CALLBACKS ###########################
-@app.callback(
-    Output("feed-content", "children"),
-    [
-        Input("left-tabs-styled-with-inline", "value"),
-        Input("intermediate-value", "children"),
-    ],
-)
-def feed_tab_content(tab_value, state):
-    """Callback to change between news and twitter feed
-    """
-    # print(f"feed tab value {tab_value}")
-    # print(f"feed tab state {state}")
-    if tab_value == "twitter-tab":
-        return twitter_feed(state)
-    else:
-        return news_feed(state)
-
-
 ########################################################################
 #
 #                       Confirm/Death Table
@@ -117,84 +84,6 @@ stats_tabs = dbc.Card(
     ],
     className="stats-table-div",
 )
-
-
-@app.callback(
-    Output("stats-table", "children"), [Input("intermediate-value", "children"),],
-)
-def stats_tab_content(state):
-    df = stats_table(state)
-
-    # font_size_heading = ".4vh"
-    font_size_body = ".9vw"
-    table = dash_table.DataTable(
-        data=df.to_dict("records"),
-        columns=[
-            {"name": "State/County", "id": "State/County",},
-            {
-                "name": "Confirmed",
-                "id": "Confirmed",
-                "type": "numeric",
-                "format": Format(group=","),
-            },
-            {
-                "name": "Deaths",
-                "id": "Deaths",
-                "type": "numeric",
-                "format": Format(group=","),
-            },
-        ],
-        editable=False,
-        sort_action="native",
-        sort_mode="multi",
-        column_selectable="single",
-        style_as_list_view=True,
-        fixed_rows={"headers": True},
-        fill_width=False,
-        style_table={
-            "width": "100%",
-            "height": "100vh",
-        },
-        style_header={
-            "backgroundColor": color_bg,
-            "border": color_bg,
-            "fontWeight": "bold",
-            "font": "Lato, sans-serif",
-            "height": "2vw",
-        },
-        style_cell={
-            "font-size": font_size_body,
-            "font-family": "Lato, sans-serif",
-            "border-bottom": "0.01rem solid #313841",
-            "backgroundColor": "#010915",
-            "color": "#FEFEFE",
-            "height": "2.75vw",
-        },
-        style_cell_conditional=[
-            {
-                "if": {"column_id": "State/County",},
-                "minWidth": "4vw",
-                "width": "4vw",
-                "maxWidth": "4vw",
-            },
-            {
-                "if": {"column_id": "Confirmed",},
-                "color": "#F4B000",
-                "minWidth": "3vw",
-                "width": "3vw",
-                "maxWidth": "3vw",
-            },
-            {
-                "if": {"column_id": "Deaths",},
-                "color": "#E55465",
-                "minWidth": "3vw",
-                "width": "3vw",
-                "maxWidth": "3vw",
-            },
-        ],
-    )
-
-    return table
 
 
 ########################################################################
@@ -244,22 +133,6 @@ us_maps_tabs = dbc.Card(
         ]
     ),
 )
-
-
-@app.callback(
-    Output("us-map", "figure"),
-    [
-        Input("middle-map-tabs-styled-with-inline", "value"),
-        Input("intermediate-value", "children"),
-    ],
-)
-def map_tab_content(value, state):
-    """Callback to change between news and twitter feed
-    """
-    if value == "testing-us-map-tab":
-        return drive_thru_scatter_mapbox(state=REVERSE_STATES_MAP[state])
-    else:
-        return confirmed_scatter_mapbox(state=REVERSE_STATES_MAP[state])
 
 
 ########################################################################
@@ -424,102 +297,3 @@ desktop_body = [
         ]
     ),
 ]
-
-
-########################################################################
-#
-#                           Confirm cases callback
-#
-########################################################################
-@app.callback(
-    [Output("confirmed-cases-timeline", "figure")],
-    [Input("intermediate-value", "children")],
-)
-def confirmed_cases_callback(state):
-    fig = cases_chart(state)
-    return [fig]
-
-
-@app.callback(
-    [Output("confirmed-cases-chart-title", "children")],
-    [Input("intermediate-value", "children")],
-)
-def confirmed_cases_callback(state="US"):
-    if state == "US":
-        return ["U.S. Confirmed Cases"]
-    else:
-        return [f"{REVERSE_STATES_MAP[state]} Confirmed Cases"]
-
-
-########################################################################
-#
-#                           Deaths callback
-#
-########################################################################
-@app.callback(
-    [Output("deaths-timeline", "figure")], [Input("intermediate-value", "children")]
-)
-def confirmed_cases_callback(state):
-    fig = deaths_chart(state)
-    return [fig]
-
-
-@app.callback(
-    [Output("death-chart-title", "children")], [Input("intermediate-value", "children")]
-)
-def death_callback(state="US"):
-    if state == "US":
-        return ["U.S. Deaths"]
-    else:
-        return [f"{REVERSE_STATES_MAP[state]} Deaths"]
-
-
-########################################################################
-#
-#                           Trajectory callback
-#
-########################################################################
-@app.callback(
-    [Output("infection-trajectory-title", "children")],
-    [Input("intermediate-value", "children")],
-)
-def trajectory_title_callback(state="US"):
-    if state == "US":
-        return ["U.S. Trajectory"]
-    else:
-        return [f"{REVERSE_STATES_MAP[state]} Trajectory"]
-
-
-@app.callback(
-    [Output("infection-trajectory-chart", "figure")],
-    [Input("intermediate-value", "children")],
-)
-def trajectory_chart_callback(state):
-    fig = new_infection_trajectory_chart(state)
-    return [fig]
-
-
-########################################################################
-#
-#                           Top bar callback
-#
-########################################################################
-@app.callback(
-    [Output("daily-stats", "children")], [Input("intermediate-value", "children")]
-)
-def daily_stats_callback(state):
-    cards = daily_stats(state)
-    return [cards]
-
-
-########################################################################
-#
-#                   State Dropdown Menu Callback
-#
-########################################################################
-@app.callback(
-    [Output("intermediate-value", "children")], [Input("states-dropdown", "value")]
-)
-def update_output(state):
-    state = STATES_COORD[state]["stateAbbr"]
-    return [state]
