+++ ncov19_dash/wsgi.py
@@ -1,6 +1,6 @@
 from werkzeug.middleware.dispatcher import DispatcherMiddleware
-from .app import server
-from .app import app
+# from .app import server
+# from .app import app
 
 application = DispatcherMiddleware(server, {
     '/': app.server,
